+++ CHANGELOG.md
@@ -5,9 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.7.1] - 2018-07-20
+
 ### Added
 
 - better error for invalid query selector strings
+
+### Fixed
+
+- a bug where parent interactors were returned within nested
+  methods when using deeper nested methods
 
 ## [0.7.0] - 2018-07-06